--- a/src/Parlot/Fluent/AndSkip.cs
+++ b/src/Parlot/Fluent/AndSkip.cs
@@ -19,6 +19,8 @@ namespace Parlot.Fluent
         {
             context.EnterParser(this);
 
+            var start = context.Scanner.Cursor.Position;
+
             if (_parser1.Parse(context, ref result))
             {
                 if (_parser2.Parse(context, ref _parseResult2))
@@ -26,7 +28,7 @@ namespace Parlot.Fluent
                     return true;
                 }
 
-                context.Scanner.Cursor.ResetPosition(result.Start);
+                context.Scanner.Cursor.ResetPosition(start);
             }
 
             return false;