--- a/src/Parlot/Fluent/AndSkip.cs
+++ b/src/Parlot/Fluent/AndSkip.cs
@@ -2,14 +2,14 @@
 
 namespace Parlot.Fluent
 {
-    public sealed class AndSkip<T> : Parser<T>
+    public sealed class AndSkip<T, U> : Parser<T>
     {
         internal readonly IParser<T> _parser1;
-        internal readonly IParser _parser2;
+        internal readonly IParser<U> _parser2;
 
-        static ParseResult<object> _parseResult2 = new ParseResult<object>();
+        static ParseResult<U> _parseResult2 = new ParseResult<U>();
 
-        public AndSkip(IParser<T> parser1, IParser parser2)
+        public AndSkip(IParser<T> parser1, IParser<U> parser2)
         {
             _parser1 = parser1 ?? throw new ArgumentNullException(nameof(parser1));
             _parser2 = parser2 ?? throw new ArgumentNullException(nameof(parser2));