--- a/src/Parlot/Fluent/Between.cs
+++ b/src/Parlot/Fluent/Between.cs
@@ -2,11 +2,11 @@
 
 namespace Parlot.Fluent
 {
-    public sealed class Between<T> : Parser<T>
+    public sealed class Between<A, T, B> : Parser<T>
     {
         private readonly IParser<T> _parser;
-        private readonly IParser _before;
-        private readonly IParser _after;
+        private readonly IParser<A> _before;
+        private readonly IParser<B> _after;
 
         private readonly bool _beforeIsChar;
         private readonly char _beforeChar;
@@ -16,7 +16,7 @@ namespace Parlot.Fluent
         private readonly char _afterChar;
         private readonly bool _afterSkipWhiteSpace;
 
-        public Between(IParser before, IParser<T> parser, IParser after)
+        public Between(IParser<A> before, IParser<T> parser, IParser<B> after)
         {
             _before = before ?? throw new ArgumentNullException(nameof(before));
             _parser = parser ?? throw new ArgumentNullException(nameof(parser));
@@ -41,8 +41,6 @@ namespace Parlot.Fluent
         {
             context.EnterParser(this);
 
-            var parsed = new ParseResult<object>();
-
             if (_beforeIsChar)
             {
                 if (_beforeSkipWhiteSpace)
@@ -57,7 +55,9 @@ namespace Parlot.Fluent
             }
             else
             {
-                if (!_before.Parse(context, ref parsed))
+                var parsedA = new ParseResult<A>();
+
+                if (!_before.Parse(context, ref parsedA))
                 {
                     return false;
                 }
@@ -82,7 +82,9 @@ namespace Parlot.Fluent
             }
             else
             {
-                if (!_after.Parse(context, ref parsed))
+                var parsedB = new ParseResult<B>();
+
+                if (!_after.Parse(context, ref parsedB))
                 {
                     return false;
                 }