--- a/src/Parlot/Fluent/DecimalLiteral.cs
+++ b/src/Parlot/Fluent/DecimalLiteral.cs
@@ -23,7 +23,7 @@ namespace Parlot.Fluent
                 context.SkipWhiteSpace();
             }
 
-            var start = context.Scanner.Cursor.Position;
+            var start = context.Scanner.Cursor.Offset;
 
             if ((_numberOptions & NumberOptions.AllowSign) == NumberOptions.AllowSign)
             {
@@ -36,9 +36,9 @@ namespace Parlot.Fluent
 
             if (context.Scanner.ReadDecimal())
             {
-                var end = context.Scanner.Cursor.Position;
+                var end = context.Scanner.Cursor.Offset;
 
-                if (decimal.TryParse(context.Scanner.Buffer.AsSpan(start.Offset, end - start), NumberStyles.AllowLeadingSign | NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture, out var value))
+                if (decimal.TryParse(context.Scanner.Buffer.AsSpan(start, end - start), NumberStyles.AllowLeadingSign | NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture, out var value))
                 {
                     result.Set(context.Scanner.Buffer, start, end, Name, value);
                     return true;