--- a/src/Parlot/Fluent/DecimalLiteral.cs
+++ b/src/Parlot/Fluent/DecimalLiteral.cs
@@ -40,7 +40,7 @@ namespace Parlot.Fluent
 
                 if (decimal.TryParse(context.Scanner.Buffer.AsSpan(start, end - start), NumberStyles.AllowLeadingSign | NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture, out var value))
                 {
-                    result.Set(context.Scanner.Buffer, start, end, Name, value);
+                    result.Set(start, end,  value);
                     return true;
                 }
             }