--- a/src/Parlot/Fluent/Identifier.cs
+++ b/src/Parlot/Fluent/Identifier.cs
@@ -44,7 +44,7 @@ namespace Parlot.Fluent
 
             var end = context.Scanner.Cursor.Offset;
 
-            result.Set(context.Scanner.Buffer, start, end, Name, new TextSpan(context.Scanner.Buffer, start, end - start));
+            result.Set(start, end, new TextSpan(context.Scanner.Buffer, start, end - start));
             return true;
         }
     }