--- a/src/Parlot/Fluent/IntegerLiteral.cs
+++ b/src/Parlot/Fluent/IntegerLiteral.cs
@@ -40,7 +40,7 @@ namespace Parlot.Fluent
 
                 if (long.TryParse(context.Scanner.Buffer.AsSpan(start, end - start), NumberStyles.AllowLeadingSign | NumberStyles.AllowDecimalPoint, CultureInfo.InvariantCulture, out var value))
                 {
-                    result.Set(context.Scanner.Buffer, start, end, Name, value);
+                    result.Set(start, end,  value);
                     return true;
                 }
             }