--- a/src/Parlot/Fluent/OneOrMany.cs
+++ b/src/Parlot/Fluent/OneOrMany.cs
@@ -35,7 +35,7 @@ namespace Parlot.Fluent
 
             } while (_parser.Parse(context, ref parsed));
 
-            result = new ParseResult<List<T>>(context.Scanner.Buffer, start, end, Name, results);
+            result = new ParseResult<List<T>>(start, end, results);
             return true;
         }
     }