--- a/src/Parlot/Fluent/ParseResult.cs
+++ b/src/Parlot/Fluent/ParseResult.cs
@@ -1,33 +1,23 @@
 ﻿namespace Parlot
 {
-    public interface IParseResult
+    public struct ParseResult<T>
     {
-    }
-
-    public struct ParseResult<T> : IParseResult
-    {
-        public ParseResult(string buffer, int start, int end, string parserName, T value)
+        public ParseResult(int start, int end, T value)
         {
-            Buffer = buffer;
             Start = start;
             End = end;
             Value = value;
-            ParserName = parserName;
         }
 
-        public void Set(string buffer, int start, int end, string parserName, T value)
+        public void Set(int start, int end, T value)
         {
-            Buffer = buffer;
             Start = start;
             End = end;
             Value = value;
-            ParserName = parserName;
         }
 
         public int Start;
         public int End;
-        public string Buffer;
         public T Value;
-        public string ParserName;
     }
 }