… src/Parlot/Fluent/StringLiteral.cs
… src/Parlot/Fluent/Switch.cs
… src/Parlot/Fluent/TextBefore.cs
… src/Parlot/Fluent/TextLiteral.cs
… src/Parlot/Fluent/Then.cs
… src/Parlot/Fluent/WhiteSpaceLiteral.cs
… src/Parlot/Fluent/ZeroOrMany.cs
… src/Parlot/ITokenResult.cs
… src/Parlot/Scanner.cs
… src/Parlot/TokenResult.cs