Removed some `TextPosition` and replaced by `int` when possible.
Strongly-typed some parsers to prevent some conversion allocations and intermediate `Parse<object>` callse

Using Fluid benchmarks:

Before

|         Method |      Mean |    Error |   StdDev |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|--------------- |----------:|---------:|---------:|--------:|--------:|--------:|----------:|
|          Parse |  11.04 us | 0.189 us | 0.158 us |  0.7477 |       - |       - |    3.1 KB |

After

|         Method |       Mean |      Error |     StdDev |      Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|--------------- |-----------:|-----------:|-----------:|--------:|--------:|--------:|----------:|
|          Parse |   8.534 us |  0.1417 us |  0.1326 us |    0.7019 |       - |       - |   2.87 KB |