@sebastienros 2021-01-07
Bonus, it's now faster than the "hand made" Scriban parser

|         Method |         Mean |      Error |     StdDev |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|--------------- |-------------:|-----------:|-----------:|--------:|--------:|--------:|----------:|
|          Parse |     9.120 us |  0.1843 us |  0.1724 us |  1.0071 |       - |       - |   4.16 KB |