@lahma 2021-01-07
Looking good. I tried for the kicks to cope without `IParser` and most trouble came from the untyped access, this probably would help with that too if ever considered.