@sebastienros 2021-01-07
@lahma yes. I am desperately looking for a way to remove `IParser`. At least the part that proxies the actual logic to the `Parser<T>` implementations when using untyped parsers. Tried to have it not return any result for example to prevent allocations. Assuming that calling an untyped parser means we don't care about the result. But it's not always true ...